fe: extracted header content of Column into its own component
