+++ frontend/cypress/integration/retro/board.spec.js
@@ -76,11 +76,5 @@ context("Board Tests", () => {
       .contains("Save")
       .click();
     bp.shouldHaveColumnTitleWith("Mad-edited");
-
-    cy.pause();
-  });
-
-  it("should edit card", () => {
-    // TODO: implement
   });
 });
